backend/requirements_production.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
gTTS==2.5.0
requests==2.31.0
python-multipart==0.0.6
email-validator==2.1.0

# Production-specific
psycopg2-binary==2.9.9
python-decouple==3.8
cloudinary==1.36.0
gunicorn==21.2.0
